Fixes #4834 - suggested edit for Date.getTime doc
As noted in #4834, the documentation for this API uses unspecific language, making this API feel very unreliable. Here's the proposed update. @simn has noted milestone 3.4 on the issue, but I don't see why improving the docs should wait for a release.

Here's the proposal - take it, leave it, or discuss further. :)